+++ src/lib/components/v2/admin/UserSettingsEditor.svelte
@@ -5,7 +5,7 @@
 
 	/** @type {string} */
 	export let runnerBackend;
-	/** @type {import('$lib/types').UserSettings} */
+	/** @type {import('fractal-components/types/api').UserSettings} */
 	export let settings;
 	/** @type {string} */
 	export let settingsApiEndpoint;
@@ -14,7 +14,7 @@
 
 	let settingsFormSubmitted = false;
 
-	/** @type {import('$lib/types').UserSettings} */
+	/** @type {import('fractal-components/types/api').UserSettings} */
 	let originalSettings;
 
 	export let pendingChanges = false;
@@ -204,7 +204,7 @@
 				</div>
 			</div>
 		{/if}
-		{#if runnerBackend !== 'local' && runnerBackend !== 'local_experimental'}
+		{#if runnerBackend !== 'local'}
 			<div class="row mb-3 has-validation">
 				<label for="slurmAccount-0" class="col-sm-3 col-form-label text-end">
 					<strong>SLURM accounts</strong>