+++ src/lib/components/v2/workflow/VersionUpdateFixArgs.svelte
@@ -5,7 +5,7 @@
 
 	/** @type {import('fractal-components/types/api').WorkflowTaskV2} */
 	export let workflowTask;
-	/** @type {import('$lib/types').Task|import('fractal-components/types/api').TaskV2} */
+	/** @type {import('fractal-components/types/api').TaskV2} */
 	export let updateCandidate;
 	/** @type {boolean} */
 	export let parallel;