+++ src/routes/v2/tasks/management/+page.server.js
@@ -4,7 +4,7 @@ import { listTaskGroups } from '$lib/server/api/v2/task_api';
 
 export async function load({ fetch }) {
 	const user =
-		/** @type {import('$lib/types').User & {group_ids_names: Array<[number, string]>}} */ (
+		/** @type {import('fractal-components/types/api').User & {group_ids_names: Array<[number, string]>}} */ (
 			await getCurrentUser(fetch, true)
 		);
 	const taskGroups = removeIdenticalTaskGroups(await listTaskGroups(fetch), user);